feat: log trainingdataset scraper, chore: deprecate scripts that will be part of data_gen and training pipeline, feat: first scraping data in v0.3.1
