+++ logDataset/dataset_gen_pipeline/subtools/source.py
@@ -3,6 +3,7 @@ import json
 import csv
 import shutil
 import requests
+import html
 
 from helpers import helpers
 from subtools.subtool import Subtool
@@ -44,11 +45,12 @@ class SourceSubtool(Subtool):
                                  "Exception", "exception", "error", "Error", "warn",
                                    "Warn", "Warning", "signal" , "INFO", "info", "Info",
                                    "exit", "code", "Code", "CODE", "traceback", "Traceback"]
+        body_logs_excluders = ["{", "}", "\t", "if", "def", "func", "class", "=", ";"]
         log_markdown_begin = ["\r\n```\r\n", "`"]
         log_markdown_end = ["\r\n```\r\n", "`"]
         query = "https://api.stackexchange.com/2.3/search/advanced?pagesize=25&page={}&order=desc&sort=creation&answers=1&site=stackoverflow&filter=!*236eb_eL9rai)MOSNZ-6D3Q6ZKb0buI*IVotWaTb&body={}"
         discovered_logs = []
-        pages = 25
+        pages = 2
         page = 1
         for keyword in body_log_keywords:
             page = 1
@@ -56,20 +58,27 @@ class SourceSubtool(Subtool):
                 response = requests.get(query.format(page, keyword))
                 if response.status_code == 200:
                     response = json.loads(response.text)
-                    for item in response["items"]:
+                    items_num = len(response["items"])
+                    for itemid, item in enumerate(response["items"]):
                         body = item["body_markdown"]
+                        print('Processing {} out of {} items'.format(itemid+1, items_num))
                         for lidx, log_begin in enumerate(log_markdown_begin):
                             while log_begin in body:
                                 body_idx_start_seq = body.find(log_begin) + len(log_begin)
                                 body_idx_end_seq = body[body_idx_start_seq:].find(log_markdown_end[lidx])
                                 log = body[body_idx_start_seq:body_idx_start_seq + body_idx_end_seq]
-                                if any(validator in log for validator in body_logs_validators):
+                                if (any(validator in log for validator in body_logs_validators) and 
+                                    all(log.count(excluder) < 5 for excluder in body_logs_excluders)):
                                     discovered_logs.append(log)
+                                    body = body.replace(log_begin, "", 1)
+                                    body = body.replace(log_markdown_end[lidx], "", 1)
                                 else:
                                     body = body.replace(log_begin, "", 1)
                                     body = body.replace(log_markdown_end[lidx], "", 1)
+                response = None
+                print('Processed {} page out of {}'.format(page, pages))
                 page += 1
-        self.__pushLogs(discovered_logs)
+        self.__pushLogs(discovered_logs, batch_size_def=1)
                     
                         
     def __scrapeGithub(self):
@@ -110,15 +119,15 @@ class SourceSubtool(Subtool):
             f.close()
         self.__pushLogs(logs)
     
-    def __pushLogs(self, logs):
-        batch_size=5
+    def __pushLogs(self, logs, batch_size_def=5):
         log_batch_string = ""
+        batch_size = batch_size_def
         log_batch = []
         for log in logs:
             if batch_size <= 0:
                 log_batch.append(log_batch_string)
                 log_batch_string = ""
-                batch_size = 5
+                batch_size = batch_size_def
             batch_size -= 1
             log_batch_string += log + "\n"
         if log_batch_string != "":
@@ -130,10 +139,10 @@ class SourceSubtool(Subtool):
         new_file  = file_template.replace('v', str(new_version))
         sources_path = '../sources/'
         shutil.copyfile(sources_path + current_file + '.csv', sources_path + new_file + '.csv')
-        with open(sources_path + new_file + '.csv', 'a') as f:
+        with open(sources_path + new_file + '.csv', 'a', encoding="utf-8") as f:
             csv_writer = csv.writer(f, delimiter='\n')
             for log in log_batch:
-                log = log.replace(',', ' ')
+                log = self.__logCleanup(log)
                 csv_writer.writerow([log])
     
     def __getNewSourceVersion(self, source_file_template):
@@ -144,3 +153,11 @@ class SourceSubtool(Subtool):
             return "v0.0.1"
         else:
             return files_matching_template[-1].split('_')[-1].replace('.csv', '')
+        
+    def __logCleanup(self, log):
+        log = log.replace(',', ' ')
+        log = log.replace('\n', ' ')
+        log = log.replace('\r', ' ')
+        log = log.replace('\t', ' ')
+        log = html.unescape(log)
+        return log
